% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeVarsNames.R
\name{ChangeVarsNames}
\alias{ChangeVarsNames}
\title{Helper function to automatically change several variables' names}
\usage{
ChangeVarsNames(x, hs_m = FALSE, m_log2 = FALSE, adj_log2 = FALSE)
}
\arguments{
\item{x}{A name of the variable whose name is to be changed}

\item{hs_m}{A logical indicating which name transformation to perform}

\item{m_log2}{A logical indicating which name transformation to perform}

\item{adj_log2}{A logical indicating which name transformation to perform}
}
\value{
A string with changed variable name
}
\description{
Helper function to automatically change several variables' names
}
