% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustFatCreat.R
\name{.AdjustFatCreat}
\alias{.AdjustFatCreat}
\title{Adjust exposure variables concentrations with creatinine concentration or total fat percentage}
\usage{
.AdjustFatCreat(data, list_fat_adj, list_creat_adj)
}
\arguments{
\item{data}{A dataframe containing variables to be adjusted}

\item{list_fat_adj}{A character vector with names of variables to be adjusted for total fat percentage}

\item{list_creat_adj}{A character vector with names of variables to be adjusted for creatinine}
}
\value{
A dataframe containing original and adjusted variables
}
\description{
Adjust exposure variables concentrations with creatinine concentration or total fat percentage
}
