% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogTransform.R
\name{.LogTransform}
\alias{.LogTransform}
\title{log transforms exposure variables concentrations}
\usage{
.LogTransform(data, list_log2, list_ln)
}
\arguments{
\item{data}{A dataframe containing variables to be log transformed}

\item{list_log2}{A character vector with names of variables to be log2 transformed}

\item{list_ln}{A character vector with names of variables to be ln transformed}
}
\value{
A dataframe containing original and transformed variables
}
\description{
log transforms exposure variables concentrations
}
