% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IqrTransform.R
\name{IqrTransform}
\alias{IqrTransform}
\title{IQR standardize exposure variables concentrations}
\usage{
IqrTransform(dataset)
}
\arguments{
\item{dataset}{An imputed dataset (mids object or a dataframe)}
}
\value{
A dataframe with IQR transformed variables replacing original ones
}
\description{
IQR standardize exposure variables concentrations
}
