% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareDatasets.R
\name{PrepareDatasets}
\alias{PrepareDatasets}
\title{Prepares all raw datasets for further processing}
\usage{
PrepareDatasets(
  exposure_data,
  covariate_data,
  covariates_list,
  sex_age_data,
  fat_adjust,
  creatinine_adjust,
  log2_transform,
  ln_transform
)
}
\arguments{
\item{exposure_data}{A .csv file containing maternal exposures (data.frame, rows = subjects, cols = exposure variables)}

\item{covariate_data}{An .Rdata file containing covariates (data.frame, rows = subjects, cols = covariate variables)}

\item{covariates_list}{A character vector containing covariates names}

\item{sex_age_data}{A .csv file containing child age and sex covariates (data.frame, rows = subjects, cols = variables containing sex and age)}

\item{fat_adjust}{A character vector of names of variables to be adjusted with fat}

\item{creatinine_adjust}{A character vector of names of variables to be adjusted with creatinine}

\item{log2_transform}{A character vector of names of variables to be log2 transformed}

\item{ln_transform}{A character vector of names of variables to be ln transformed}
}
\value{
A dataframe with all outcomes, exposures and covariates of interest
}
\description{
Prepares all raw datasets for further processing
}
